requirements.txt
streamlit
pandas
numpy
matplotlib
scipy
python-dotenv
google-genai
